requirements.txt
absl-py==2.1.0
accelerate==1.0.1
addict==2.4.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
anytree==2.12.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
arxiv==2.1.3
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==5.0.1
attrs==23.2.0
Babel==2.15.0
backcall==0.2.0
beautifulsoup4==4.12.3
bert-score==0.3.13
bioc==2.1
bitsandbytes==0.45.0
bleach==6.1.0
blinker==1.8.2
boto3==1.35.79
botocore==1.35.79
Brotli==1.1.0
cachetools==5.5.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
conllu==4.5.3
contourpy==1.1.1
cos_python_sdk_v5==1.9.36
crcmod==1.7
cycler==0.12.1
Cython==3.0.11
datasets==3.1.0
DateTime==5.5
debugpy==1.8.2
decorator==5.1.1
deepspeed==0.15.2
defusedxml==0.7.1
Deprecated==1.2.15
dill==0.3.8
distro==1.9.0
docopt==0.6.2
duckduckgo_search==5.3.1b1
editdistance==0.8.1
einops==0.8.0
et_xmlfile==2.0.0
exceptiongroup==1.2.1
executing==2.0.1
fastjsonschema==2.20.0
feedparser==6.0.11
filelock==3.16.1
flair==0.14.0
fonttools==4.55.3
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.9.0
ftfy==6.2.3
func_timeout==4.3.5
fuzzywuzzy==0.18.0
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.37.0
google-auth-oauthlib==1.0.0
griffe==1.4.0
grpcio==1.68.1
h11==0.14.0
h2==4.1.0
h5py==3.11.0
hjson==3.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.26.5
hyperframe==6.0.1
icecream==2.1.3
idna==3.7
imageio==2.35.1
importlib_metadata==8.0.0
importlib_resources==6.4.0
intervaltree==3.1.0
ipykernel==6.29.5
ipython==8.12.3
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.1
jieba==0.42.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
json5==0.9.25
jsonlines==4.0.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter-server==1.24.0
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server_terminals==0.5.3
jupyterlab==4.3.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
jupyterlab_widgets==3.0.13
kafka-python==2.0.2
kiwisolver==1.4.7
lagent==0.2.3
langdetect==1.0.9
language_tool_python==2.8.1
lazy_loader==0.4
lemminflect==0.2.3
lru-dict==1.3.0
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mmengine==0.10.5
more-itertools==10.5.0
mpld3==0.5.10
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
narwhals==1.17.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.1
ninja==1.11.1.2
nltk==3.9.1
notebook==7.3.1
notebook_shim==0.2.4
num2words==0.5.13
numpy==1.24.4
oauthlib==3.2.2
omegaconf==2.3.0
openai==1.3.5
opencv-python==4.10.0.84
OpenHowNet==2.0
openpyxl==3.1.5
overrides==7.7.0
packaging==24.1
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.4
peft==0.13.2
pexpect==4.9.0
phx-class-registry==4.0.6
pickleshare==0.7.5
pillow==10.4.0
pinyin==0.4.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.2
portalocker==3.0.0
pptree==3.1
prettytable==3.11.0
progressbar2==4.5.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
propcache==0.2.0
protobuf==5.29.1
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycocoevalcap==1.2
pycocotools==2.0.7
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.10.3
pydantic_core==2.27.1
pydeck==0.9.1
Pygments==2.18.0
pynvml==11.5.3
pyparsing==3.1.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-utils==3.8.2
pytorch_revgrad==0.2.0
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.19.0
rsa==4.9
s3transfer==0.10.4
safetensors==0.4.5
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
segtok==1.5.11
semver==3.0.2
Send2Trash==1.8.3
sentencepiece==0.2.0
sgmllib3k==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
socksio==1.0.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlitedict==2.1.0
stack-data==0.6.3
streamlit==1.40.1
sty==1.0.6
supervisor==4.2.5
sympy==1.13.3
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
terminado==0.18.1
terminaltables==3.1.10
textattack==0.3.10
threadpoolctl==3.5.0
tifffile==2023.7.10
tiktoken==0.7.0
timeout-decorator==0.5.0
timm==0.9.10
tinycss2==1.3.0
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
torch==2.1.1+cu118
torchaudio==2.1.1+cu118
torchvision==0.16.1+cu118
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
transformer-smaller-training-vocab==0.4.0
transformers==4.44.2
transformers-stream-generator==0.0.5
triton==2.1.0
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.2
uri-template==1.3.0
urllib3==1.26.20
urwid==3.0.0
uvicorn==0.15.0
uvloop==0.21.0
validators==0.34.0
visual-genome==1.1.1
watchdog==4.0.2
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.6
widgetsnbextension==4.0.13
Wikipedia-API==0.7.1
word2number==1.1
wrapt==1.17.0
XlsxWriter==3.2.0
xmltodict==0.14.2
xtuner==0.1.23
xxhash==3.5.0
yapf==0.43.0
yarl==1.15.2
zipp==3.19.2
zope.interface==7.2
